made stuff work lol
