+++ src/components/map.js
@@ -1,12 +1,16 @@
 import React from 'react';
-import { StyleSheet, View } from 'react-native';
+import { StyleSheet, View, Image } from 'react-native';
 import MapboxGL from '@react-native-mapbox-gl/maps';
-import PeerLocationsView from './peerLocations'
+import { connect } from 'react-redux';
+
+const IMAGE_PIGGY = require("../assets/images/piggy.png")
+
 
 MapboxGL.setAccessToken('pk.eyJ1IjoidGVzc29yby0iLCJhIjoiY2t1b3EzY2d2MGV1ejJ2bzFtbXIxMmdjbCJ9.P-4uyej3lnQVxNs1Tzc-Sw');
 
 
-export default MapView = () => {
+const MapView = (props) => {
+  console.log("Mapview update")
   return (
     <View style={styles.map}>
       <MapboxGL.MapView
@@ -26,8 +30,21 @@ export default MapView = () => {
         </MapboxGL.Camera>
         <MapboxGL.UserLocation />
 
+        {props.locations.map(peerLocation => {
+          return (<MapboxGL.MarkerView
+            id={peerLocation.id}
+            key={peerLocation.id}
+            coordinate={[peerLocation.location.Longitude, peerLocation.location.Latitude]}>
+            <View style={styles.iconWrapper}>
+              <Image
+                source={IMAGE_PIGGY}
+                style={styles.icon}
+              />
+            </View>
+          </MapboxGL.MarkerView>)
+        })}
 
-        <PeerLocationsView />
+
 
       </MapboxGL.MapView>
     </View>
@@ -40,4 +57,25 @@ const styles = StyleSheet.create({
     height: '100%',
     width: '100%',
   },
+  icon: {
+    height: 15,
+    width: 15,
+    opacity: 1,
+    resizeMode: 'contain',
+    borderRadius: 10
+  },
+  iconWrapper: {
+    backgroundColor: 'rgb(255, 179, 0)',
+    padding: 5,
+    borderRadius: 20,
+    borderWidth: 2,
+    borderColor: 'rgba(255,255,255,0.7)'
+  }
 });
+
+
+const mapStateToProps = (state) => {
+  return { locations: state.peerLocations }
+};
+const connectComponent = connect(mapStateToProps);
+export default connectComponent(MapView);
